requirements.txt
django==4.2.*
djangorestframework
mysqlclient
python-dotenv
djangorestframework-simplejwt
drf-yasg
django-cors-headers